+++ blog-portal/src/main/java/site/iblogs/portal/service/CommentService.java
@@ -5,6 +5,6 @@ import site.iblogs.portal.model.request.CommentRequest;
 import site.iblogs.portal.model.response.CommentResponse;
 
 public interface CommentService {
-    PageResponse<CommentResponse> getComment(int cid,int pageNum, int pageSize);
+    PageResponse<CommentResponse> getComment(Long cid,int pageNum, int pageSize);
     Boolean SaveComment(CommentRequest comment);
 }
